Feat: 메인 페이지 데이터 연결 및 장바구니, 모달 디자인 약간 수정
