+++ src/components/common/Header/Header.tsx
@@ -163,7 +163,7 @@ const Header = () => {
         <StyledHeaderWhiteBox>
           <StyledHeaderWhiteContent>
             <div>
-              <BsArrowLeft size="40" onClick={handleArrowLeft} />
+              <StyledLeftBtn size="40" onClick={handleArrowLeft} />
             </div>
             <div onClick={handleReservationText}>
               {location.pathname === '/cart'
@@ -183,7 +183,7 @@ const Header = () => {
           <StyledHeaderContent>
             {location.pathname === '/regionList' ? (
               <StyledHeaderRegionCover>
-                <BsArrowLeft size="40" onClick={handleArrowLeft} />
+                <StyledLeftBtn size="40" onClick={handleArrowLeft} />
                 <StyledHeaderRegion>{selectedRegion}</StyledHeaderRegion>
                 <StyledChevronDown
                   onClick={() => {
@@ -397,6 +397,10 @@ const StyledChevronDown = styled(ChevronDown)`
   height: 1.5rem;
   fill: ${theme.colors.blue};
   margin-left: 0.5rem;
+  cursor: pointer;
+`;
+
+const StyledLeftBtn = styled(BsArrowLeft)`
   cursor: pointer;
 `;
 export default Header;
