+++ src/pages/main/components/mainAllListContainer.tsx
@@ -1,6 +1,6 @@
-import { MainListProps } from './getPlaces';
 import MainAllListItem from './mainAllListItem';
 import { StyledTitle, StyledWrapper } from './mainListContainer';
+import { MainListProps } from './mainListItem';
 
 const MainAllListContainer = ({ title }: MainListProps) => {
   return (
