+++ src/pages/main/components/mainListItem.tsx
@@ -1,13 +1,7 @@
 import { theme } from '@styles/theme';
 import styled from 'styled-components';
-import { FaStar } from 'react-icons/fa';
 import { useQuery } from '@tanstack/react-query';
-import {
-  MainListProps,
-  MainItemProps,
-  getMostSell,
-  getFavorite,
-} from './getPlaces';
+import { getMostSell, getFavorite } from './getPlaces';
 import Slider from 'react-slick';
 import 'slick-carousel/slick/slick.css';
 import 'slick-carousel/slick/slick-theme.css';
@@ -15,12 +9,32 @@ import { eclipsText } from '@utils/textLength';
 import { useNavigate } from 'react-router-dom';
 import { formatDate } from '@utils/useFormatDate';
 
+export interface MainListProps {
+  title: string;
+  data?: {
+    body: MainItemProps[];
+  };
+}
+
+export interface MainItemProps {
+  page: number;
+  size: number;
+  id: number;
+  name: string;
+  latitude: number;
+  longitude: number;
+  addressCode: string;
+  likeCount: number;
+  phoneNumber: string;
+  accommodationImageUrl: string;
+  price: number;
+}
+
 const MainListItem = ({ title }: MainListProps) => {
   const navigate = useNavigate();
   const { data } = useQuery<MainListProps>({
     queryKey: [title],
     queryFn: title === '많이 판매된 숙소' ? getMostSell : getFavorite,
-    // refetchInterval: 1000,
   });
   const settings = {
     infinite: true,
@@ -67,7 +81,6 @@ const MainListItem = ({ title }: MainListProps) => {
                 {item.price.toLocaleString('ko-KR')}원~
               </StyledItemPrice>
             </StyledItemDesc>
-            {item.like ? <StyledStar /> : <StyledNoStar />}
           </StyledMainPageItem>
         ))}
     </StyledWrapper>
@@ -101,7 +114,7 @@ const StyledMainPageItem = styled.div`
   width: 9.75rem;
   height: 10.5rem;
   border-radius: 1rem;
-  box-shadow: ${theme.shadows.shadow1.shadow};
+  box-shadow: ${theme.shadows.shadow2.shadow};
   position: relative;
   cursor: pointer;
 `;
@@ -127,15 +140,4 @@ const StyledItemName = styled.p`
 
 const StyledItemPrice = styled.p`
   font-size: 0.8rem;
-`;
-
-export const StyledStar = styled(FaStar)`
-  color: ${theme.colors.yellow};
-  position: absolute;
-  top: 7.5rem;
-  right: 0.2rem;
-`;
-
-const StyledNoStar = styled(StyledStar)`
-  color: ${theme.colors.gray2};
 `;
